+++ server/src/ffmpeg/builder/MediaStream.ts
@@ -1,5 +1,6 @@
 import type { ExcludeByValueType, Nullable } from '@/types/util.js';
-import { isEmpty, isNull, merge } from 'lodash-es';
+import { isNonEmptyString } from '@tunarr/shared/util';
+import { isEmpty, isNull, merge, nth } from 'lodash-es';
 import type { AnyFunction, MarkOptional, StrictOmit } from 'ts-essentials';
 import { VideoFormats } from './constants.ts';
 import { PixelFormatUnknown, type PixelFormat } from './format/PixelFormat.ts';
@@ -126,6 +127,29 @@ export class VideoStream implements MediaStream {
 
     const [num, den] = this.sampleAspectRatio.split(':');
     return parseFloat(num) / parseFloat(den);
+  }
+
+  getNumericFrameRateOrDefault(defaultRate: number = 24) {
+    let frameRate = defaultRate;
+    if (isNonEmptyString(this.frameRate)) {
+      const intParsed = parseInt(this.frameRate);
+      if (isNaN(intParsed)) {
+        const parts = this.frameRate.split('/');
+        const numerator = nth(parts, 0);
+        const denominator = nth(parts, 1);
+        let rate = 24;
+        if (isNonEmptyString(numerator) && isNonEmptyString(denominator)) {
+          const numeratorInt = parseInt(numerator);
+          const denominatorInt = parseInt(denominator);
+          if (!isNaN(numeratorInt) && !isNaN(denominatorInt)) {
+            rate = Math.round(numeratorInt / denominatorInt);
+          }
+        }
+        frameRate = rate;
+      }
+    }
+
+    return frameRate;
   }
 }