+++ server/src/ffmpeg/ffmpegInfo.ts
@@ -315,8 +315,6 @@ export class FfmpegInfo {
         '-show_format',
         '-show_chapters',
         '-show_streams',
-        '-analyzeduration',
-        '30',
         `${path}`,
       ],
       { timeout, swallowError: false },