+++ web/src/components/settings/media_source/PlexServerEditDialog.tsx
@@ -58,7 +58,6 @@ const emptyDefaults: PlexServerSettingsForm = {
   uri: '',
   name: '',
   accessToken: '',
-  sendChannelUpdates: false,
   sendGuideUpdates: false,
   index: 0,
   type: 'plex',
@@ -334,20 +333,6 @@ export function PlexServerEditDialog({ open, onClose, server }: Props) {
                     />
                   }
                   label="Auto-Update Guide"
-                />
-              </FormControl>
-              <FormControl>
-                <FormControlLabel
-                  control={
-                    <Controller
-                      control={control}
-                      name="sendChannelUpdates"
-                      render={({ field }) => (
-                        <Checkbox {...field} checked={field.value} />
-                      )}
-                    />
-                  }
-                  label="Auto-Update Channels"
                 />
               </FormControl>
             </Box>